UTF-8 encoding, tag list -> text file
Used unicode() funtion to include utf-8 encoding. This works fine and didn't observe any errors being thrown.
Unique tag list is now written to output file whose value is taken from args.outpath. However it is to be noted that the tag list is appended to the existing logs in the out put file